UIButton設定を別ファイルへ移動
